Refactoring of the db interface: One db for position and event information only. Trip id removed from that db.

+++ test/src/com/ja/saillog/test/android/TestExportFile.java
@@ -83,7 +83,9 @@ public class TestExportFile extends AndroidTestCase {
                                             false);
         }
         finally {
-            f.delete();
+            if (null != f) {
+                f.delete();
+            }
         }
     }
